Semantic fixes to make warp pass solidity's semantic tests squashed:
fix uint256_byte

fix Or in BuiltinHandler.py

fix signextend

fix uint256_sar

partial sha fix

Fix sar

fix ForLoopEliminator: 'leave' now makes all returns to be parameters

If there is a 'leave' in a function, all of its return variables, must
     be either assigned to before that point, or be also parameters of
     the function. It wasn't the case before this commit. Since
     checking that all the variables has been assigned to is kind of
     complex, we chose an easy way for now and just require all return
     variables to also be parameters.

fix the order of arguments in u256_shr application

Fix for loops

If there is a continue in the for loop the post conditions were being
skipped. The fix was to move the conditions to just before the recursive
loop call.

Fix typos

adjust keccak to conform to 'unsafe_keccak' expectations

Problem: the standard 'unsafe_keccak' function expects that the the
     last number in the data array is big-endian, but leading zeroes are
     truncated, so the resulting size is less than 16 bytes. We
     however don't truncate them and that triggers an assert in
     'unsafe_keccak'.

Solution: truncate the last number in the data array for 'unsafe_keccak'.

fix the order of arguments in `uint256_bytes`…

fix extraction of if-blocks

There were several issues in ScopeFlattener.

1. The function '_is_leave' had a wrong condition for ifs. An if node
     is 'leave'ing iff each branch is leaving. If there is no
     else-branch — if is non-leaving.

2. Extractor of ifs has a check, if the if can be inlined. It can be
     inlined if one branch of the if is terminating — either leaving
     or reverting. However, it checks the node _before_ the subsequent
     transformation, but returns the node _after_ the
     transformation. Sometimes the transformation makes a branch
     non-terminating (well, it can be established that it is
     terminating through control flow analysis, but the Cairo compiler
     can't figure it out). So, we check for the terminating property
    _after_ the transformation.

3. Another way to fix this issue above, and also another optimization,
     is to _help_ the compiler with the control flow analysis. If we
     _know_ that a certain branch is leaving, after extraction, we
     don't add a check for leaving but rather just add a leave
     instruction.

Co-authored-by: Artem Yurchenko <artemyurchenko@zoho.com>
Co-authored-by: BitBaseBit <vardygreg23@gmail.com>
Co-authored-by: Jorik Schellekens <joriksch@gmail.com>