Transpile "argless" fallback functions (#334)
* Add cairo writing of "argless" fallback functions

Update decorators writer, so functions can have multiple decorators
at the same time.

If name is of kind `fallback` name it as `__default__`

* Add compilation test for argless fallback function

* Modify `astChecking.ts` cond for constructors

Change condition to check for function kind instead of an empty name since
fallback function have empty names as well

* Update `behaviour.test.ts` find function

* Minor fix

* Add behaviour tests for fallback functions

* Minor improvement

* Fix constructor checking condition

* Modify `identifierMangler` to not modify nameless functions

* Check for empty func names on special funcs

Check that a function is nameless if and only if it has kind
`constructor`, `fallback` or `receive`.